Create NewSecretModel
